+++ event-details-boundary/Event.ts
@@ -1,10 +1,7 @@
-import { placemarkToFormattedAddress } from "@lib/AddressFormatting"
-import { setStringAsync as expoSetClipboardStringAsync } from "expo-clipboard"
-import { showLocation } from "react-native-map-link"
 import { EventArrival, EventArrivalsTracker } from "@arrival-tracking"
 import { CurrentUserEvent } from "@shared-models/Event"
 import { dayjs } from "TiFShared/lib/Dayjs"
-import { EventLocation } from "TiFShared/domain-models/Event"
+import { isAttendingEvent } from "TiFShared/domain-models/Event"
 
 /**
  * A type for the color value for an event.
@@ -17,67 +14,6 @@ export enum EventColors {
   Pink = "#F7B2BD",
   Orange = "#F4845F",
   Yellow = "#F6BD60"
-}
-
-export type EventLocationCoordinatePlacemark = Pick<
-  EventLocation,
-  "coordinate" | "placemark"
->
-
-/**
- * Copies an event location to the clipboard.
- *
- * If the event has no formattable placemark, the coordinates are formatted
- * and copied to the clipboard instead of the formatted address of the placemark.
- */
-export const copyEventLocationToClipboard = (
-  location: EventLocationCoordinatePlacemark,
-  setClipboardText: (text: string) => Promise<void> = expoCopyTextToClipboard
-) => setClipboardText(formatEventLocation(location))
-
-const expoCopyTextToClipboard = async (text: string) => {
-  await expoSetClipboardStringAsync(text)
-}
-
-const formatEventLocation = (location: EventLocationCoordinatePlacemark) => {
-  const formattedLocationCoordinate = `${location.coordinate.latitude}, ${location.coordinate.longitude}`
-  if (!location.placemark) return formattedLocationCoordinate
-  const formattedPlacemark = placemarkToFormattedAddress(location.placemark)
-  return formattedPlacemark ?? formattedLocationCoordinate
-}
-
-/**
- * Opens the event location in the user's preffered maps app.
- *
- * @param location the location to open in maps.
- * @param directionsMode the mehtod of how to link specific directions to the location.
- */
-export const openEventLocationInMaps = (
-  location: EventLocationCoordinatePlacemark,
-  directionsMode?: "car" | "walk" | "bike" | "public-transport"
-) => {
-  showLocation({
-    ...location.coordinate,
-    title: location.placemark
-      ? placemarkToFormattedAddress(location.placemark)
-      : undefined,
-    directionsMode
-  })
-}
-
-/**
- * Returns true if the status indicates that the user is hosting the event.
- */
-export const isHostingEvent = (attendeeStatus: EventUserAttendeeStatus) => {
-  return attendeeStatus === "hosting"
-}
-
-/**
- * Returns true if the status indicates that the user is either hosting or
- * attending the event.
- */
-export const isAttendingEvent = (attendeeStatus: EventUserAttendeeStatus) => {
-  return attendeeStatus !== "not-participating"
 }
 
 /**
